Ikke slett deltakerlister med deltakere på. (#396)
* Hendelse i prod har medført at en enkeltplass gjennomføring er slettet selv om det er en deltaker på. Dette er mest sannsynlig fordi valp ikke bryr seg om visse gjennomføringsstatuser men vi kan ikke slette deltakerene fordi vi får ikke ny melding fra arena etter at gjennomføringen evt er gjenopprettet

+++ amt-deltaker/src/main/kotlin/no/nav/amt/deltaker/kafka/GjennomforingConsumer.kt
@@ -47,8 +47,15 @@ class GjennomforingConsumer(
         value: String?,
     ) {
         if (value == null) {
-            if (key !in deltakerlisteTombstoneBlacklist) {
+            log.info("Mottok tombstone for gjennomføring: $key")
+            val antallDeltakere = deltakerRepository.getAntallDeltakereForDeltakerliste(key)
+            if (key !in deltakerlisteTombstoneBlacklist && antallDeltakere == 0) {
                 deltakerlisteRepository.delete(key)
+            } else {
+                log.error(
+                    "Ignorerer tombstone for $key. " +
+                        "Deltakerliste er svartelistet eller har deltakere. Antall deltakere: $antallDeltakere",
+                )
             }
         } else {
             handterDeltakerliste(objectMapper.readValue(value))
